rename ipfsApi to ipfs
